[c++] Cap max bytes read in GenericReadVariableUnsigned
GenericReadVariableUnsigned will keep reading as long as the high bit is
set, regardless of the width of the var int being read. This can cause
it to consume too many bytes when given an invalid encoding for a var
int.

Now, GenericReadVariableUnsigned will throw an StreamException when it
detects such cases.

The specialized InputBuffer::ReadVariableUnsigned already has such a
limit in place in the recursive template VariableUnsignedUnchecked.